requirements.txt
Django==2.1.5
psycopg2>=2.7,<3.0
psycopg2-binary
djangorestframework==3.9.1
Pillow==5.4.1
selenium==3.141.0